+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [5.4.0] - 2025-10-07
+- Added `FilePickerShell` component
+- Added 4 new `Button` types: `glass`, `glassDark`, `dangerGlass`, and `selectedGlass`
+- Added `flat` prop to `Button`
+- Updated dependencies
+
 ## [5.3.3] - 2025-09-25
 - Added `event.stopPropagation` by default on `MenuItem`s. This can be overridden with the new `onClickNative` prop, which receives the native click event.
 - Fixed `ToggleButton` styles.
@@ -422,6 +428,7 @@ Co-authored with @piqusy
 - Initial release
 
 [Unreleased]: https://github.com/infinum/eightshift-ui-components/compare/master...HEAD
+[5.4.0]: https://github.com/infinum/eightshift-ui-components/compare/5.3.3...5.4.0
 [5.3.3]: https://github.com/infinum/eightshift-ui-components/compare/5.3.2...5.3.3
 [5.3.2]: https://github.com/infinum/eightshift-ui-components/compare/5.3.1...5.3.2
 [5.3.1]: https://github.com/infinum/eightshift-ui-components/compare/5.3.0...5.3.1